feat: extract cache from eager loader
